requirements.txt
streamlit==1.42.0
numpy==2.2.2
langchain_ollama==0.2.3
langhchain_core==0.3.34
faiss-cpu
serpapi==0.1.5
Wikipedia-API==0.8.1
python-dotenv==1.0.1